Changed folder structure to organize by robot type instead of task type.
