+++ source/robotis_lab/robotis_lab/tasks/manager_based/OMY/cabinet/cabinet_env_cfg.py
