+++ source/robotis_lab/robotis_lab/tasks/manager_based/OMY/stack/config/agents/robomimic/bc_rnn_image_84.json
