--- source/robotis_lab/robotis_lab/tasks/manager_based/__init__.py
@@ -1,28 +1,0 @@
-# Copyright 2025 ROBOTIS CO., LTD.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-# Author: Taehyeong Kim
-
-# Copyright (c) 2022-2025, The Isaac Lab Project Developers.
-# All rights reserved.
-#
-# SPDX-License-Identifier: BSD-3-Clause
-
-"""Manipulation environments for fixed-arm robots."""
-
-import gymnasium as gym
-from .cabinet import *  # noqa
-from .lift import *  # noqa
-from .reach import *  # noqa
-from .stack import *  # noqa
